refactor: split CreatePlanModal component
