+++ packages/arex/src/panes/Replay/AppTopBar/RecordedCase/RecordedCaseDetail.tsx
